dynamic descriptor heap

+++ LaughingEngine/LaughingEngine/GameTimer.cpp
@@ -1,8 +1,8 @@
 #include "PCH.h"
 #include "GameTimer.h"
 
-double GameTimer::m_SecondsPerCount = 0.0f;
-double GameTimer::m_DeltaTime = 0.0f;
+float GameTimer::m_SecondsPerCount = 0.0f;
+float GameTimer::m_DeltaTime = 0.0f;
 
 int64_t GameTimer::m_BaseTime = 0;
 int64_t GameTimer::m_PausedTime = 0;
@@ -16,10 +16,10 @@ void GameTimer::Initialize()
 {
 	LARGE_INTEGER frequency;
 	QueryPerformanceFrequency(&frequency);
-	m_SecondsPerCount = 1.0 / static_cast<double>(frequency.QuadPart);
+	m_SecondsPerCount = 1.0f / (float)frequency.QuadPart;
 }
 
-double GameTimer::TotalTime()
+float GameTimer::TotalTime()
 {
 	if (m_Stopped)
 	{
@@ -31,7 +31,7 @@ double GameTimer::TotalTime()
 	}
 }
 
-double GameTimer::DeltaTime()
+float GameTimer::DeltaTime()
 {
 	return m_DeltaTime;
 }
@@ -108,12 +108,12 @@ int64_t GameTimer::GetCurrentTick()
 	return static_cast<int64_t>(currentTick.QuadPart);
 }
 
-double GameTimer::DeltaTimeMillisecs(int64_t tick1, int64_t tick2)
+float GameTimer::DeltaTimeMillisecs(int64_t tick1, int64_t tick2)
 {
-	return (tick2 - tick1) * m_SecondsPerCount * 1000.0;
+	return (tick2 - tick1) * m_SecondsPerCount * 1000.0f;
 }
 
-double GameTimer::DeltaTimeSeconds(int64_t tick1, int64_t tick2)
+float GameTimer::DeltaTimeSeconds(int64_t tick1, int64_t tick2)
 {
 	return (tick2 - tick1) * m_SecondsPerCount;
 }
